move `treestatus_ui.py` to `views/ui.py`
